Move artifact download code to separate file
This will help avoid mutually-recursive `require`s.
